feat: ドメイン駆動設計に基づくプロパティメソッドAPIの実装
ADR-006に基づき、回帰分析APIをドメイン駆動設計の原則に従って再設計:

- 新しいプロパティメソッドAPI: slope(), intercept(), r_value(), p_value(), stderr(), intercept_stderr()
- 既存get_params()メソッドに非推奨警告を追加（後方互換性維持）
- ファイル構造をモジュール化して保守性を向上
- 全24テストが成功

変更されたファイル:
- rustgression/regression/: 機能ごとに分割（base.py, ols.py, tls.py, factory.py）
- tests/: テスト責務別に分割（test_base.py, test_ols.py, test_tls.py, test_comparison.py）
